requirements.txt
Flask==2.2.2
line-bot-sdk==2.3.0
requests==2.28.1
gunicorn==20.1.0